Fix compatibility with newer hubot-slack adapter
In hubot-slack v4.2.2 msg.topic has been renamed to msg.setTopic, to
avoid collision with events, it brakes the topic setting.
Check if the adapter has setTopic method and call it the same way hubot
would call.
If no such method, fall back to the previous behavior, and call the
topic method.